requirements.txt
json5
trio
eventlet
python-socketio[client]
